requirements.txt
langchain
langchain-community
pillow
screen-brightness-control
pydantic
SQLAlchemy
psutil
pycaw
ttkbootstrap 